fix: review migration

+++ apps/banking_api/priv/repo/migrations/20210415132403_create_users_table.exs
@@ -7,7 +7,7 @@ defmodule BankingApi.Repo.Migrations.CreateUsersTable do
       add :name, :string
       add :email, :string
       add :password, :string
-      add :balance, :integer
+      add :balance, :bigint
 
       timestamps()
     end
